review pt2
